+++ student_application/lib/screens/project-email-screen.dart
@@ -1,13 +1,11 @@
-
 import 'package:flutter/material.dart';
 
 class ProjectEmailScreen extends StatelessWidget {
   const ProjectEmailScreen({super.key});
+  static const routeName = '/email';
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Text('PROJECT EMAIL PAGE')
-      );
+    return Container(child: Text('PROJECT EMAIL PAGE'));
   }
 }
